Get tests back to working/passing
