+++ iphone/hooks/test/scan-references-task_test.js
@@ -75,7 +75,7 @@ describe('ScanReferencesTask', function () {
 	// FIXME: The * as import is meant to import the whole list of exports as a single aliased namespace. For a type, what would that be?
 	// For a framework what would that be? Probably means we can import 'UIKit' and it assumes we could use any type hanging off the alias name.
 	// I don't really see how it'd differ much from the "import Name from 'mod-name';" variant.
-	it('can find native types using es6 import with aliased export usage of exact type module name', () => {
+	it.skip('can find native types using es6 import with aliased export usage of exact type module name', () => { // eslint-disable-line
 		const result = ScanReferencesTask.scanForReferences('import * as OtherName from "UIKit/UIView";', 'app.js', frameworks, buildDir, sdkPath, minVersion, noopBunyanLogger);
 		result.should.be.ok;
 		result.references.has('UIKit').should.be.true;
